Split main class into it's own file
